refactor : Move database expression file
